requirements.txt
pygame==2.5.2
transformers 
torch
pandas
openai 
azure-cognitiveservices-speech